+++ pkg/R/miscellaneous.R
@@ -146,7 +146,7 @@ gen.primes <- function(n = 1, domain = 2:n, random = FALSE, distinct = TRUE, cha
 	# gen.primes(n = 100, domain = c(450, 9571), random = FALSE, distinct = TRUE) %>% { data.table::data.table(x = ., d = c(0, diff(.))) } %T>% print %>% plot
 }
 
-call.recursion <- function(x, fun, test, nxt, max.iter = 1, cur.iter = 0, simplify = TRUE){
+call.recursion <- function(x, fun, test, ..., nxt = I, max.iter = 1, cur.iter = 0, simplify = TRUE){
 	#' Execute a Function as Recursively
 	#'
 	#' \code{call.recursion} Executes a recursive calls to \code{fun} based on the output of \code{test} up to \code{max.iter} times
@@ -154,6 +154,7 @@ call.recursion <- function(x, fun, test, nxt, max.iter = 1, cur.iter = 0, simpli
 	#' @param x The input object
 	#' @param fun A function that operates on \code{x} and produces output
 	#' @param test A single-argument function returning a single Boolean: \code{FALSE} stops iteration
+	#' @param ... Additional arguments passed to \code{fun}
 	#' @param nxt A function that operates on the current output of \code{fun(x)} to send to the \emph{next} iterative call to \code{fun()}
 	#' @param max.iter (integer) The maximum number of iterations
 	#' @param cur.iter (integer) The current iteration index
@@ -180,7 +181,7 @@ call.recursion <- function(x, fun, test, nxt, max.iter = 1, cur.iter = 0, simpli
 
 		x <- nxt(ans);
 
-		ans <- fun(x);
+		ans <- fun(x, ...);
 
 		iterations[[cur.iter]] <- mget(c("cur.iter", "x", "ans")) |> purrr::modify_at(c("x", "ans"), list);
 	}